refactor: create folders for matches and no matches components
